chore(tests): put control tests in river
