+++ tests/river/unit/common/test_db_connection.py
