+++ tests/river/unit/common/test_fetch_mapping.py
